requirements.txt
requests
python-dotenv
selenium
webdriver-manager
beautifulsoup4
pytz
openai 